requirements.txt
requests
numpy
pandas
scikit-learn
streamlit
matplotlib
plotly
reportlab
python-dotenv
langchain
langgraph
groq
